@@ -1,1 +1,1 @@
-2.0.0-p247
+ruby-2.2.0
