@@ -12,8 +12,8 @@ module JasperRails
       if Mime::Type.lookup_by_extension(file_extension).nil?
         Mime::Type.register options[:mime_type], file_extension.to_sym
       end
-      if  Rails.version > "4.1"
-        define_render_rails_greater_4_1(renderer,file_extension, options)
+      if  Rails.version >= '4.2'
+        define_render_rails_geq_4_2(renderer,file_extension, options)
       else
         define_render_rails_less_4_2(renderer,file_extension, options)
       end
@@ -23,51 +23,38 @@ module JasperRails
       self.after_fill_blocks += [block]
     end
 
-    def self.define_render_rails_less_4_2(renderer,file_extension, opts = {})
+    def self.define_render_rails_geq_4_2(renderer, file_extension, options = {})
+      ActionController::Renderers.add "#{file_extension}".to_sym do |resource, opts|
+        JasperRails::AbstractRenderer.response_report(self, renderer, resource, options.merge(opts))
+      end
+    end
+
+    def self.define_render_rails_less_4_2(renderer,file_extension, options = {})
       ActionController::Responder.send(:define_method, "to_#{file_extension}") do
-        if template_name = self.options[:template]
-        else
-          template_name = "#{controller.controller_path}/#{controller.action_name}"
+        JasperRails::AbstractRenderer.response_report(controller, renderer, resource, options.merge(self.options))
+      end
+    end
+
+     def self.response_report(controller, renderer, resource, options)        
+        template = "#{controller.controller_path}/#{controller.action_name}"
+
+        if !options[:template].blank? && options[:template] != controller.action_name
+          template = options[:template]
         end
+        
         details_options = {:formats => [:jrxml], :handlers => []}
 
-        jrxml_file = controller.lookup_context.find_template(template_name,[],false,[],details_options).identifier
-        jasper_file = jrxml_file.sub(/\.jrxml$/, ".jasper")      
-  
+        jrxml_file = controller.lookup_context.find_template(template,[],false,[], details_options).identifier
+        jasper_file = jrxml_file.sub(/\.jrxml$/, ".jasper")
+
         params = {}
         controller.instance_variables.each do |v|
           params[v.to_s[1..-1]] = controller.instance_variable_get(v)
         end
   
-        response_options = JasperRails.config[:response_options].merge(:type => opts[:mime_type])
-        controller.send_data renderer.render(jasper_file, resource, params, self.options), response_options
-      end
-    end
-
-    def self.define_render_rails_greater_4_1(renderer,file_extension, options = {})
-      ActionController::Renderers.add "#{file_extension}".to_sym do |resource,opts|
-        default_template = "#{self.controller_path}/#{self.action_name}"
-        if opts[:template] != self.action_name
-          template_name = opts[:template]
-        else
-          template_name = default_template
-        end
-        
-        details_options = {:formats => [:jrxml], :handlers => []}
-
-        jrxml_file = self.lookup_context.find_template(template_name,[],false,[],details_options).identifier
-        jasper_file = jrxml_file.sub(/\.jrxml$/, ".jasper")      
-  
-        params = {}
-        self.instance_variables.each do |v|
-          params[v.to_s[1..-1]] = self.instance_variable_get(v)
-        end
-        
         response_options = JasperRails.config[:response_options].merge(:type => options[:mime_type])
-        self.send_data renderer.render(jasper_file, resource, params, opts), response_options
+        
+        controller.send_data renderer.render(jasper_file, resource, params, options), response_options
       end
-    end
-  
   end
-
 end
